rename tempMenu to menu
