+++ src/data/types.ts
@@ -1,3 +1,3 @@
-import type Menu from './tempMenu';
+import type Menu from './menu';
 
 export type SkuId = keyof typeof Menu;
